+++ assets/javascripts/kitten/components/navigation/header-nav/components/burger-menu.js
@@ -23,6 +23,7 @@ const ButtonIcon = props => (
 )
 
 const buttonClassNames = classNames(
+  'k-HeaderNav__BurgerMenu__button',
   'k-ButtonIcon',
   'k-ButtonIcon--tiny',
 )